refactor: Rename connected-account module to account-manager
* Rename and relocate account connection logic to a more generic account-manager module

    * refactor: Rename `:data:connected-account` modules to `:data:account-manager` across api, implementation, and testing layers
    * refactor: Rename `ConnectedAccountRepository` interface to `AccountManager` and `ConnectedProvider` to `AccountProvider`
    * refactor: Update package names to `com.thomaskioko.tvmaniac.accountmanager`
    * refactor: Rename and update implementation to `DefaultAccountManager` and `FakeAccountManager`
    * refactor: Update `ProviderScoped` utility and its `getActiveProvider` extension to use the new `AccountManager` types
    * chore: Update `settings.gradle.kts` and module `build.gradle.kts` files to reflect the new module structure
