+++ data/account-manager/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/accountmanager/api/AccountProvider.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.accountmanager.api
+
+public enum class AccountProvider {
+    TRAKT,
+    SIMKL,
+}
