+++ data/account-manager/implementation/build.gradle.kts
@@ -1,0 +1,26 @@
+plugins {
+    alias(libs.plugins.app.kmp)
+}
+
+scaffold {
+    useMetro()
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(projects.data.accountManager.api)
+                api(projects.data.traktauth.api)
+                api(libs.coroutines.core)
+            }
+        }
+
+        commonTest {
+            dependencies {
+                implementation(libs.bundles.unittest)
+                implementation(projects.data.traktauth.testing)
+            }
+        }
+    }
+}
