+++ data/account-manager/testing/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    alias(libs.plugins.app.kmp)
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(projects.data.accountManager.api)
+                api(libs.coroutines.core)
+            }
+        }
+    }
+}
