--- data/connected-account/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/connectedaccount/api/ConnectedAccountRepository.kt
@@ -1,10 +1,0 @@
-package com.thomaskioko.tvmaniac.connectedaccount.api
-
-import kotlinx.coroutines.flow.Flow
-
-public interface ConnectedAccountRepository {
-    public val activeProvider: Flow<ConnectedProvider?>
-    public val isConnected: Flow<Boolean>
-    public val connectionEvents: Flow<ConnectedProvider>
-    public fun getActiveProvider(): ConnectedProvider?
-}
