--- data/connected-account/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/connectedaccount/api/ConnectedProvider.kt
@@ -1,6 +1,0 @@
-package com.thomaskioko.tvmaniac.connectedaccount.api
-
-public enum class ConnectedProvider {
-    TRAKT,
-    SIMKL,
-}
