--- data/connected-account/implementation/build.gradle.kts
@@ -1,26 +1,0 @@
-plugins {
-    alias(libs.plugins.app.kmp)
-}
-
-scaffold {
-    useMetro()
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.data.connectedAccount.api)
-                api(projects.data.traktauth.api)
-                api(libs.coroutines.core)
-            }
-        }
-
-        commonTest {
-            dependencies {
-                implementation(libs.bundles.unittest)
-                implementation(projects.data.traktauth.testing)
-            }
-        }
-    }
-}
