+++ settings.gradle.kts
@@ -139,9 +139,9 @@ include(
     ":data:cast:api",
     ":data:cast:implementation",
     ":data:cast:testing",
-    ":data:connected-account:api",
-    ":data:connected-account:implementation",
-    ":data:connected-account:testing",
+    ":data:account-manager:api",
+    ":data:account-manager:implementation",
+    ":data:account-manager:testing",
     ":data:database:sqldelight",
     ":data:database:testing",
     ":data:datastore:api",
